sftp-s3-backup-tool/requirements.txt
paramiko
boto3
python-dotenv